Actually launch and stop the product.
Invoke the functions in Hermes to launch and stop the executables.
Add a JSON config file which contains executables and parameters to
launch the parts of the product.  I put this here instead of in
Hermes because maybe different kinds of tests will want to launch it differently.

Testing:
- Kicked off the test run, with a ten second pause once everything
was launched.
- Used ps to verify that processes were up.
- Ran a few curl commands and received valid responses.
- When the test case ended, verified that all of the processes were stopped.