Add support to launch the product.
Added a function to expand shell variables so that the location "~" can be used
in the config file.
Added functions to launch, and later stop, all executables that are part of the
product: 4 replicas, 3 clients, and 1 api server.

Testing: Ran make build run_all_tests.  See the matching athena_testing change
for details.