fix the protobuf build in docker
For some reason, when building in docker, the plugin definition is not
enough to pull in the rebar protocol buffer compiler properly. I've
added it to the "dev" profile to be pulled in as a standard dep. It is
not needed in the prod profile, because it doesn't run at release
time.

Also added erlang-parsetools to apt install, because erlang_protobuffs
relies on leex from that package.