Add a way to intentionally skip test cases.
Investigating and tweaking the framework to handle these few remaining tests
will be nontrivial.  I'm setting some aside for now and focusing on having the
framework handle more types of tests.  This change adds a way for test suites to
skip test cases.  A test suite can have a file under suites/skipped to define
these.  For the CoreVMTests suite, it is a JSON file with the format path:reason.

If the framework decides a test cannot be run at run time (in other words, it is
not in the skipped file list, but has to be skipped anyway), it will be written
to a file called unintentionally_skipped_tests.  This is the file that used to be
called skipped_tests.