fix first-round protobuf compilation
Before this change, when you ran "mkdir build && cd build && cmake
..", you got an error about being unable to find
build/athena.pb.cc. The file was generated in src/, and if you went
ahead and ran "make", it would move to build/, and you wouldn't see
this error any more.

This change fixes the issue by defining an explicit build target for
the proto definition (the "proto" library) and attaching a public
include directory specification to it. The proto file has also been
moved to its own directory, which allows the executables in tools/ to
express their dependency on it correctly.

The specification for building athena is also moved to a
CMakeLists.txt in the src/ directory, to separate those details more
cleanly. This means the athena executable is now in build/src/athena
instead of just build/athena.