fix protobuf parsing
Passing a char* to ParseFromString calls the
std::string::string(char*) constructor, which creates a string that
stops at the first zero in the string, like a C string would
have.

This PR calls ParseFromArray instead, so we can pass the correct
length of the message. It also checks the result of the function, to
find out if we parsed it correctly.