Add support for passing data to a contract invocation.
Several instructions, such as CALLDATACOPY, CALLDATASIZE, and CALLDATALOAD,
require data to be passed to the contract. This change looks for a "data" field
in the test case, and if found:

1. Creates the contract.
2. Invokes the contract with the passed in data.

If the test case does not have a "data" field, then this change has no effect.
The code will run when the contract is created, and no other time.

This change also add skips for three tests which rely on setting initial gas.