add genesis.json to docker build
The default athena.config specifies a genesis_block. The alternative
was to comment that out and have athena run with an empty genesis
block, but let's exercise as much code as we can.

The docker build uses the example genesis block from the test
directory. The balances were larger than a uint64_t, which caused an
exception to be thrown from stoull. I filed #34 to handle the
exception better, but this patch also brings the values in this file
down to a usable range.