use copy assignment instead of memcpy for hash
With the move to use evm_uint256be instead of std::vector<uint8_t> for
hashes, we can just copy them to fields using '=' instead of memcpy or
std::copy out of a vector.

Also don't have to take the address of the first element of the
internal byte array (which use to invoke the & operator for the
vector).