toward const correctness
Marked a bunch of arguments as const, because they don't need to be
modified. Marked a bunch of methods as const, because they don't
modify the object.

Had to swith the array-style square bracket map accessors to find()
with iterator checking, because the square brackets may modify the
maps, but I think that probably demonstrated a bug we didn't know
about.