Look at expected results in test cases' "out" field.
Some tests' expected results are in a field called "out"; this is the value
returned by running the code. Up until now, we've only looked at the values
which test cases put into storage.

This is done by creating a second contract which contains the bytecode to call
the code in the contract under test.

This change does not include passing parameters to the bytecode under test; that
will be a subsequent change.