fix tx order and tx leak
This started as a fix to Amit's note that a bug existed: if an
eth_call created child transactions, those transactions wouldn't be
recorded until the next eth_sendTransaction came in. We now think that
we should block child transactions for eth_call executions, but this
change is still a good idea, and a clean way to deal with the current
issue.

Two additional fixes came out of that work:

1. Transactions were in the wrong order in the block. They are now in
start-of-evaluation order.

2. Failed contract creations were not recorded. They are now.