Adding stubs for net_version and eth_accounts.
Truffle's test utility calls net_version and eth_accounts when it starts up.
In order to run the PowerCoin tests against helen/athena, these have to be
implemented.

net_version: We aren't sure if/how we'll use it in VMware's blockchain solution.
For now, it is hard coded in athena and the plumbing is in place for the value
to be retrieved from athena through helen to the caller.

eth_accounts: Truffle uses the first account retrieved by an eth_accounts call
as the "from" to upload contracts. We are still figuring out the details of user
implementation. For now, return an array with one hard coded account.

This change also adds a maven plugin to allow Java compiler arguments via
-DcompilerArgument. This lets us see deprecated usage messages, for example.
Sample usage:

mvn -DcompilerArgument=-Xlint:deprecation clean install package

Requires an athena change to be merged first:
https://github.com/vmwathena/athena/pull/62