Encapsulate RLP creation
Remove the duplicated bit mashing code in the three functions that
create RLP encodings, and define an RLPBuilder class to do it for us.

Also fixed a bug: long lists were adding the list length to the
length-length byte, instead of the length-length (a.k.a
field_length_count).

Resolved #51.