hack for geth/web3 race bug
geth's web3.js contains this code:

    var filter = contract._eth.filter('latest', function(e){
    ...
                filter.stopWatching(function() {});

This is a race condition. The callback function may be executed before
the result of contract._eth.filter is bound to the variable filter. If
that happens, the filter in the callback will be undefined.

This happens all the time when connecting geth to Helen/Athena. Since
we can return a new block immediately, if there is nothing else
happening on the geth node, we will return a result immediately,
causing the JS VM to immediately call the callback instead of
finishing the outer function. This has been the bug keeping our
contracts from getting their addresses, and causing geth to continue
pinging the filter forever.

The real fix will be to submit a patch to geth to fix their race
condition. But in the meantime, the fix implemented here is to force a
user to ask twice before getting the first set of filter updates. New
filters are put in a "delay" list and moved out of that list when
their first query is processed. That first query returns "nothing" and
does not modify the filter state. Subsequent queries work as expected.