Added support for eth_getBlockByHash (#71)
Athena will handle API requests (blocks/{N}) and the
eth_getBlockByHash requests in the same fashion : by returning
detailed information about every transaction present in the
specified block (in addition to block details).

Helen will figure out what to pass on to the user based on the
kind of request.