KVB Integration: fix corrupted transaction keys
ReplicaImp::fetchBlockData was creating a KeyValuePair out of the raw
char pointers instead of the slices. This caused new Slices to be
created implicitly because there are single-argument Slice
constructors that accept a char*. Those constructors use strlen to
determine the size of the slice, so the length pulled from the block
header is ignored.

Unfortunately, we can't just make the Slice(char*) and Slice(string)
constructors explicit, because they are used elsewhere purposely
implicitly, such as in the Status class. So for now this error is
likely to be repeated.