KVB Integration: fix filters
Filters need to read block data, so they need to go through the
KVB/SBFT flow to get access to the KVB storage.

This change removes the direct reference from the FilterManager to the
EVM, as well as the direct link from the EVM to the FilterManager. See
vmwathena/athena#86 - I think the FilterManager functionality should
probably be moved out of Athena and into Helen.