KVB Integration: fix eth_call
To finish the first pass of KVB Integration, I disabled the eth_call
API. This change re-enables it.

I've forced its processing over SBFT's read-only channel. This has the
benefit of implicitly tracking a state we've been meaning to track
anyway: whether or not an execution is allowed to change
state. eth_sendTransaction can change state (move value, write
storage, etc.), while eth_call cannot. We knew the state in normal
execution, but if a contract contained a CALL opcode, we lost that
information. We have it again now, in the form of whether or not an
IBlocksAppender is available.

I've also removed the "pending" transactions list. That was added when
with thought CALL opcodes generated their own transactions. We haven't
needed it since we fixed that. A "pending" key/value set will come
back into play at some point, to keep track of in-flight changes
before commit, but best to do that fresh instead of trying to shoehorn
it into this old pending list.