KVB Integration: move KVBStorage pointer to athena_context
Setting (and clearing, for sanity) the txctx_kvbStorage member
variable during EVM executions was prone to error. This change puts
the pointer to the storage in athena_context.

Moving the pointer also allowed a great deal of deduplication to
happen: those callbacks can access the storage interfaces directly
instead of having to call back into the EVM object. Only the "call"
function needs to call back into the EVM object now. Thus, all of the
pass-through "ath_..." callbacks are now the actuall implementations,
and their EVM::... counterparts are removed.