fix eth rpc error passing
EthDispatcher.doPost() expects the return value of dispatch() to be a
JSON string. Errors were being returned as bare strings, instead of as
JSON objects. This change wraps them in JSON again.

Before this change, you would see in the log:

2018-05-24 12:26:23 ERROR EthDispatcher:200 - Invalid request
Unexpected character (B) at position 0.
        at org.json.simple.parser.Yylex.yylex(Yylex.java:610)
        at org.json.simple.parser.JSONParser.nextToken(JSONParser.java:269)
        at org.json.simple.parser.JSONParser.parse(JSONParser.java:118)
        at org.json.simple.parser.JSONParser.parse(JSONParser.java:81)
        at org.json.simple.parser.JSONParser.parse(JSONParser.java:75)
        at Servlets.EthDispatcher.doPost(EthDispatcher.java:192)
...

And the response would say:

{"jsonprc":"2.0","id":-1,"error":{"message":"Unable to parse request"}}

After this change, the log says:

2018-05-24 12:43:26 DEBUG EthDispatcher:206 - Response: {"jsonprc":"2.0","id":11,"error":{"message":"BRYAN"}}

And the response says:

[{"jsonprc":"2.0","id":11,"error":{"message":"BRYAN"}}]

("BRYAN" being the error message I temporarily added to athena to
debug the behavior, and "11" being the actual ID I sent with the
request.)