check for error from athena before decoding response
All of the non-eth servlets expect that the response message is
available in their parseToJSON functions. Check first in BaseServlet
whether there is an ErrorResponse, and send that error back to the
client instead of calling parseToJSON.

Before this change, a failure during block lookup replied with an
invalid block JSON object like:

HTTP/1.1 200 OK
{"number":0,"size":0,"transactions":[],"parentHash":"0x","nonce":"0x","hash":"0x"}

After this change, the response is an error like:

HTTP/1.1 500 Internal Server Error
{"error":"BRYAN"}

("BRYAN" being the message I injected to test this behavior.)

The code attempts to parse basic meaning from the message, to return a
somewhat more useful HTTP status code. Not a good long-term solution,
but hopefully better than nothing.