if replica is not running, don't call byz_modify to add block
ReplicaImp::addBlockToIdleReplica allows us to write blocks only if
the replica is in idle mode. Byz_modify can only be called if the
replica is in running mode. So, when addBlockToIdleReplica calls
addBlockInternal, addBlockInternal needs to skip the call to
Byz_modify if the replica is not running.

This change allows us to move main's create_genesis_block call back to
before the replica starts, so that the genesis block gets created
correctly.