Fixes for broken tests
Changes handleGrid functions to assign translation keys instead of fetching the strings asynchronously as an object, which was causing tests to break. Uses the translate pipe in the grid component to display the translated string from its key.
Fixes a credential check on the login component. The incorrect check was causing a router navigation to happen which was unintentional in the test.

These issues were bleeding in to other tests and causing them to fail