HEL-34: correctly decode unsigned message size
Messages between Helen and Athena are prefixed with two bytes that
encode the length of the message as an unsigned integer. All of Java's
integers are signed, so when a reply is larger than 32767 bytes, Helen
reads a negative number.

We're using a 32-bit integer to store the value after it is decoded,
so the fix here is to just clobber the upper two bytes with zeros, and
we recover the unsigned value in the lower two bytes.

I've also added a check on the send side, so that Helen doesn't try to
send Athena a message if its size is larger than two bytes can encode.