release and clear result create_address after run
In 0618d17, record_transaction was moved to KVBCommandsHandler, and
altered to look always put the value of result.create_address in the
stored transaction. Unfortunately, as Amit found in 5d75789,
result.create_address is overloaded for other purposes sometimes.

This change makes the same fix as Amit's, checking whether the result
has data that needs to be released, and releasing it if so. It also
then explicitly clears result.create_address in EVM::run, because run
should never create a contract.

Before this change, if a contract execution used the overloaded
meaning of result.create_address, SBFT might get stuck retrying the
transaction. It would do this because the overloaded use is a function
pointer, and that's different at each node. That means that the hash
of the transaction would be different at each node, so all nodes would
disagree on the result. Clearing create_address causes the nodes to
once agree on the result.

This was not caught, because we didn't have a test that triggered this
case. We do now: contract_tx in HelenAPITests in Hermes.