test contract execution to tickle evmjit create_address override bug
See vmwathena/athena#122. Sending a transaction to HelloWorld.sol's
hello function causes evmjit to use result.create_address to store a
release function pointer. Running this test before the patch in Athena
causes SBFT to go into a retry loop. With the fix, the transaction
completes, and we get a receipt, as expected.