HEL-84: Implement eth_syncing RPC method
Return "false" unconditionally for now. Ethereum nodes return "false"
to say that they are up to date, or an object with
starting/current/highest block information to say that they are
behind. We don't have information about the progress of state-transfer
yet, and we go through full consensus to read the blockchain right now
anyway. So, our "node" is always up to date.