ATH-14: fix RLP encoding small values during ecrecover
The "from" address of a signed transaction was not recovered correctly
when any of value, gas, gas price, or nonce was zero, or when value
was less than 0x7f. Zeroes should have been encoded as empty strings,
and small values should have been encoded as themselves, instead of
1-length strings of themselves.

There was also a problem with v=1 in the signature that is now corrected.