Issue #83: reject transaction timestamp too far in the future
We can't reject transactions that have old timestamps, because they
might be coming to us via retries. But, we can reject timestamps "too
far" in the future, to prevent a rogue client from advancing time at
an egregious pace. That pace is configurable, but is 10 seconds ahead
(of the local system clock) by default. This is safe, because even if
a replica's local clock is running slowly, it will eventually believe
the timestamp is acceptable, and catch up to other correct replicas.