Issue #83: revert block storage version change
This was kind of useful for implementation checking, but causes far
more trouble than it's worth for actually running. Anyone that tests
with the bumped version has to wipe their rocksdb storage before they
can resume testing with a pre-bump branch.

The bump was not needed anyway. An old version of athena can read this
new block version just fine. The worst that can happen is alternating
strings of blocks with and without timestamps, as a tester switches
back and forth between versions.