ATH-56: mark evm execution as read-only
This is the fix that should have been applied for ATH-53. This tells
the EVM that state-modification is not allowed, so it doesn't call our
set-storage callback, so we can't throw the exception that was causing
the problem.

Removing the specific exception handler in favor of the general for
all cases.