Putting pieces in place for the build system.
Jenkinsfile: Invokes the build process in a shared library.  All GitHub repos
will contain this, so for realistic code maintenance there should be no build
logic in it.

src/org/vmware/Builder.groovy: Contains any logic around whether we should build.
It is still uncertain what this will do. e.g. We might look at the branch
passed in, what repo we are in, what branches the other repos have, etc...
Details will become more clear when we know how we're going to use Jenkins to
handle our three GitHub repos.
Jenkins can have files that are declarative or script. We're using script files
in any area where we may to implement logic.
Regarding the directory structure, this is related to Jenkins. We can point
Jenkins to the Hermes repo for shared libraries.  It will automatically look in
src and vars.

vars/mainBuildSteps.groovy: This is the work Ivan did.  Big pieces of the
build/test processes are present, and some things still in progress.
It is in the "vars" subdirectory because it uses the Jenkins declarative syntax,
and Jenkins looks for shared declarative pipelines in "vars".