do not give org/consortium default values in user list
ProfilesRegistryManager expects its arguments to be null
(!Optional.isPresent) if they are unspecified. Including defaultValue
here meant that they were non-null, and set to the empty string. This
caused getUsers to take the first branch of its logic, which resulted
in a NumberFormatException in getUsersInternal.

This change removes the defaultValue for the organization and
consortium parameters, so that getUsers will take its final branch
instead, if no params are specified.