Putting Jenkinsfile back to "master" before checking in.
Next, I'll add a layer of indirection to avoid this Jenkinsfile shared lib issue.
The problem is that Jenkinsfile must use @Library, which reuires a hard coded
"version" for the shared library.  However, that needs to be changed in order
to test changes to the shared library before checking in.  Then it needs to be
set to master again.

The "library" statement (not "@Library") can be given the $BRANCH_NAME variable.
But that statement cannot be used in Jenkinsfile.