change SetOfKeyValuePairs to unordered_map
SetOfKeyValuePairs being defined as unordered_set<pair<Slice, Slice>>
means that ("foo", "1") and ("foo", "2") can both exist in the set. If
those represented writes to "foo", this is bad because now there are
two different values prepared for writing, and which order they are
applied in is undefined.

This change converts SetOfKeyValuePairs to unordered_map<Slice,
Slice>, so that only one value for the key "foo" can be stored.

This fixes flakiness in the tests named vmIOandFlowOperations, which
do several writes to the same storage location in one contract
execution.