Add an option to keep the Athena DB.
I decided to just pass the entire args structure to launchProduct() instead
of several values within the args structure, so this touched all test suites.

Ran all of the suites to verify they could still launch the product.

Set --keepAthenaDB to True in the KVBlockchainTests the second time it launches
the product, because a test is specifically checking to see if data was
persisted.