build just one athena docker image, and parameterize it at launch
Instead of building four distinct docker images, one for each athena
node, but one generic docker image, and then parameterize it at
launch. Parameterization is done via docker volumes, that provide
config files in well-known places. An example parameterization is
demonstrated in `docker/docker-compose.yml`.

All parameterization config files are expected to be in one of two
places:

 * `/athena/config-public`: Config files common to all athena nodes.
 * `/athena/config-local`: Config files specific to an instance.

In the sample parameterization, `/athena/config-public` is provided by
`docker/config-public`, and `/athena/config-local` is provided by
`docker/config-athena*`.

In the public config volume, the image expects to find:

 * `*.pub`: A public SBFT configuration file.
 * `find-docker-instances.sh`: A tool for resolving hostnames in the
   public configuration file.

Each instance copies the public SBFT config file to its local config
directory, and then resolves the hostnames in it to IP addresses. This
prevents instances from stomping on each others' changes, and allows
the public volume to be mounted read-only.

In the private config volume, the image expects to find:

 * `athena.config`: The configuration file for the athena node.

The node's configuration file points to the locations of other
files. Because this is in a volume, and not part of the image, it
could be altered to point to other locations without rebuilding the
image. In the example included here, the config expects `genesis.json`
in the public config directory, and logging and private keys in the
local config directory.

I've taken this opportunity to clean up a few other things in our
docker config:

 * All replicas now use the same ports: 5458/tcp for the communication
   with helen, and 3501-3505/udp for internal SBFT. Since they're all
   on separate (virtual) hosts, this works, and should make firewall
   configuration easier.

 * A "log" volume is also included on each replica, so that we have
   some way to get to a replica's individual logs without scrubbing
   through the docker-compose output.