ATH-69: slice rewrite; now sliver
Our use of Slice is leaking memory. Slice doesn't own the pointer to
the buffer it wraps. Some of our errors can be fixed by adding a
`delete[] slice.data()` in the right places. But in other cases, the
Slice that we have is pointing to only a portion of a larger buffer,
and we have lost the pointer to that larger buffer.

This PR begins the proposal of a replacement called Sliver. Sliver
wraps the buffer it points to in a `std::shared_ptr`. Sub-slivers use
this same shared pointer, and remember their offset within the buffer
explictly (instead of implicitly, via a different pointer). The shared
pointer handles reference counting, so that we only free the buffer
once all references to it have ended.

This commit just adds the simplest implementation and a sanity
test. The next commit will replace some uses of Slice with Sliver.