ATH-69: convert storage side of KVBlockchain to use Sliver
Sliver needed a different deallocator. All buffers are created using
new[], so they need to be freed using delete[]. Changed sliver_test to
match.

Rocksdb iterators seem to not want you to deallocate the memory they
hand you. This plays poorly with the new design of Sliver. So,
fromRocksdbSliver has been replaced by copyRocksdbSliver, which makes
a copy of the data from the iterator. This should be revisited and
corrected, so that we can still gain the efficiency of not copying,
where possible.

This is the first version of this change that compiles and runs all of
the Hermes tests.