ATH-72: cleanup KVB clients when athena exits
Valgrind complains that 2+MB is leaked from the KVBClientPool
constructor. This is because it holds onto the pointers to our 4
KVBClients, which each have a 512KB buffer. The queue that we put them
in doesn't properly delete them.

This PR pops all of the clients out of the pool when Athena shuts
down, and deletes each, to properly stop the client and free its
memory.

This was not a very dangerous leak, because it did not grow over time,
and the memory was in use basically until the end of the program. But,
it's the largest leak in small tests, and it's nice to get rid of
distractions.