fix ephemeral Sliver use in ReplicaImp
If you have tried to enable DEBUG-level logging recently, you probably
saw your Athena nodes exit immediately. The source was this debug
message in ReplicaImp that tried to wrap a nullptr in a Sliver. Sliver
asserts that any pointer you pass it is non-null.

This actuall prevented a worse error, which is that when the pointer
is non-null, wrapping it in an ephemeral Sliver like this immediately
frees its memory after the log message is printed. The code that just
called this function thus receives an invalid pointer on return.

The log message was a bad habit anyway. It would print the entire
block, which is too verbose, even for debug logging. This change just
prints the value of the pointer, so you can tell whether or not it is
null.