Stop using the withDockerRegistry plugin, and start tagging releases in git.
I upgraded some plugins in an effort to reduce the number of Jenkins restarts we have.
As a result, we hit an issue in which a new version of the withDockerRegistry Jenkins
plugin fails to authenticate: https://issues.jenkins-ci.org/browse/JENKINS-38018

After reading the discussion in the bug, I think it is best to just use the CLI.

This change also starts tagging releases in git.  In support of that:
- Jenkins now fetches code with SSH instead of HTTP.  That way, the SSH keys on the
  build system are used, and we don't have to enter git credentials.
- Jenkins now configures a git identity so that git will allow the commit of the tag.
  That change will be made in an upcoming build template.  For now, the git commands
  will be run at build time.