make athena directly buildable in monorepo
This just adds an extra "../" to the search paths for evmjit and
googletest, so that if you cloned the blockchain repo to the same
place you cloned the old athena repo, you can still build athena via
direct "mkdir build && cd build && cmake .. && make".

Docker build is probably broken by this patch, because it checks out
evmjit in a different directory. The correct fix for both will be to
either pull in both of these as submodules, or require preinstallation
of both of them somewhere "global" on the system or in the container.