ATH-82,HEL-128: Implement emit_log
Contracts are allowed to emit "logs" (also known as "events"). This
helps other contracts and observers learn when "interesting" things
have happened in a contract, without having to directly inspect its
state. We haven't supported these ... until now.

Athena changes:

Logs are gathered in a vector during execution, and then added to the
transaction receipt for storage. The logs are read and added to the
transaction lookup response when requested.

Helen changes:

Replacing the bogus empty logs that were added for truffle compatibility.

Hermes changes:

The "logs" field in an ethereum_tsts spec is the keccak hash of the
RLP encoding of the logs of a transaction receipt.

This implementation requires one hack. Since we're not setting the
caller, and we're not running all of the tests, in order, the contract
address does not match the test's expectation. The contract address is
part of the log, so our hash wouldn't match the expected hash. The
hack is to replace the actual contract address with the expected
address, just for the purposes of hashing.

TODO: This change does not include the "logsBloom" bloom filters. That
will be addressed in a follow-up issue.