[VB-351] Add initial code for provision service
This patch creates the initial basis of the Concord deployment service
targetting a SDDC deployment.

Specifically, each deployment request performs the following steps:
1. Connect to SDDC, obtain all the environment parameters necessary to
deploy a Concord "housing" for 1 Concord replica. Currently, each
"housing" is 1 VM.

2. Connect to SDDC network controller and provision the necessary
blockchain network and network plumbing. Currently, the initial
implementation targets SDDC on VMC. So the network plumbing involves
talking to NSX in order to provision 2 logical networks: 1 for Concord
replicas to communicate with each other (as well as to IPs external to
the SDDC via a gateway/router), 1 for Concord replicas to talk to the
Helen controller / mothership for lifecycle management traffic and fleet
tracking.