prevent checkstyle from reviewing proto-generated files (#29)
Jenkins was not having this issue, but in my local build, checkstyle
would try to process
helen/target/generated-sources/protobuf/java/com/vmware/athena/Athena.java,
which does not follow our style guidelines. This change (specifying
the source directory) seems to be the one most recommended for
resolving this issue.