VB-141: move filters to Helen (#30)
* VB-141: move filters to Helen

This change moves the management of ethereum filters to the RPC
endpoint, instead of keeping it in Concord. There are a few reasons
for this:

1. Helen round-robins requests across Concord nodes. Filter state is
   not replicated among Concord nodes, so it's quite likely that the
   poll request for a filter would go to a different node than the
   create request did, and thus you'd get a "not found" response,
   instead of the expected updates.

2. Filter state is not recorded on the blockchain. This makes it kind
   of awkward to keep in Concord in the first place.

3. Filter state is associated with the application that created the
   filter, so it makes more sense to keep it closer to the
   application.

This implementation keeps filters in-memory. This is simply taking
advantage of the fact that filters are designed (in ethereum) as very
ephemeral constructs that are typically used for a limited amount of
time, and therefore wasting persistence time on them is not
worthwhile.

This implementation does not replicate filter state among Helen nodes,
nor does it use the shared database. This is because this RPC call
will soon be moved out to the RPC stub we've been talking about. Once
it's out there, it will not be behind any sort of load-balancer. An
application will connect directly to a particular RPC stub, and this
is where the state of that application's filters will be kept. If the
stub goes down, the application will have to reconnect to another
stub, and just as if it were talking to a geth endpoint, it will need
to recreate any filters it needs on that node.

There is no garbage collection of old filters implemented yet. This is
a TODO, and should likely be based on how many blocks behind a filter
is.

This PR adds tests for filters:

1. Hermes tests in the ExtendedRPCTests suite verify the
   application-level behavior. This includes verifying some quirks
   (like delays) that geth/web3 depend on. These tests were run
   against the pre-existing Concord-owned filter implementation, and
   then re-run against this new implemenation, to verify continuity.

2. JUnit tests in the FiltersTests suite verify some simple properties
   of the filter object and manager.

* Do not send empty requests to Concord

The new EthFilterChangesHandler was sending empty (zero-length)
requests to Concord if an application asked for changes to a
non-existent filter. Concord assumes a zero-length message is an
error, and closes the connection. Helen eventually detects this as a
timeout while trying to read from the socket.

This change makes AbstractEthRpcHandler.buildRequest return a boolean
value, which signals whether the request should actually be sent to
Concord. Effectively, this allows the handler to decide the "isLocal"
property of EthDispatcher: if buildRequest returns false, then
buildResponse is called with a null ConcordResponse object, without
sending anything to Concord, just as we have done before with
EthLocalResponseHandler, and eth_uninstallFilter handling.