VB-181: Modify existing JWT to understand consortium
Several changes.  I reorganized some of the configuration to allow me to
create a MockMvc test.  I did this to add some test for the
UserAuthenticator controller.  This was to establish a baseline for
refactoring the controller.

Refactored the controller a bit.  Also added in the Consortium field
(the point of this PR) to the JWT.

Added a test for the JWT.