VB-184: store gasLimit in block (#45)
Replaces static gasLimit set in #44.

Web3py attempts to be helpful and looks up the gasLimit of the last
block in the chain, to check whether the estimated gas usage for the
transaction it wants to submit is higher than that cap. Concord/Helen
was not including a gasLimit field, so web3py was failing.

This change adds gasLimit to eth_getBlockBy* responses. The value that
is returned is the gas limit that was given to the EVM when executing
the transaction in the block, with two small exceptions.

Exception 1: The gasLimit in the genesis block is the gasLimit that
was given in genesis.json. Note that the value has been changed in the
example gensis.json, so that it matches the default that is returned
if no gasLimit was recorded.

Exception 2: The gasLimit in a block that was generated by a
personal_newAccount is zero. The account generation feature is moving
to Helen, where it will not generate blocks, so it is not worth
spending effort to record this value.

Choosing the gasLimit for a transaction/block follows the timestamp
pattern: the value is chosen by the node that submits the transaction
to SBFT. The reason for this is the need for all replicas to choose
the same gasLimit, so that get-block responses match later. If the
gasLimit were chosen by the replica, it would be impossible to change
the gasLimit without first stopping all replicas at the same
block. Since it is chosen by the client, the gas limit can be changed
at one node at a time (or even left different on each node).