VB-183: subtask -- Create Blockchain entity.
Create a blockchain entity, and initialize it with the current
ConcordAuthorities.

I reorganzied the default user creation, and added a new component for
getting the various default profiles.  We'll need this later to support
the existing API while we transition it.

I added a number of unit tests, and also introduced an H2 in memory DB
for testing purposes.

Updated schema.sql in both Helen and Hermes.

Please note: You'll need to run the sqlscript again on cockroach to
create the new BLOCKCHAINS table.