VB-199: centralize helen-concord protobuf (#58)
* VB-199: centralize helen-concord protobuf

A new sub-project is introduced: "communication". The purpose of this
directory is to hold utilities that an app uses to talk to concord
nodes. For now, this is the protocol buffers definition. Later, it
will likely include other pieces of helen/src/.../communication. This
is all preparation for moving the Ethereum RPC handlers out of Helen
and into a standalone process.

The protobuf is removed from both helen and concord. They now depend
on the version in communication.

Helen picks up this dependency as a jar - communication must be built
and installed (as concord-communication) before building helen. A new
stage, "Build product prereqs", is added to the Jenkins process to
build that jar.

Concord runs protoc on the file directly during its build. The docker
build must now be given the top-level of this repo as the build path,
instead of the concord subdirectory, because the communication
subdirectory must be copied into the build image. The build command in
Jenkins has been updated to call docker directly, and
docker-compose.sh has been removed. The docker-compose.sh script was a
holdover from the time when we build four unique Athena images, and
was no longer needed anyway.

Instructions for building the docker images have been added to the
top-level README.