VB-203: Expose RPC URLs in members list. (#71)
* VB-203: Expose RPC URLs in members list.

In VB-202, a separate spring boot app was created to expose the
Ethereum RPC interface. This change adds the URLs for the instances of
that app to the response from /api/concord/members.

The URL list is currently configured via the ConcordRpcUrls property,
and stored in the ConcordProperties object.

The format of the ConcordRpcUrls property is a comma separated list of
<host>=<url> pairs. The <host> portion is matched to the hostname
returned from Concord in a PeerResponse message. The url is added as
"rpc_url" in the JSON response.

* doc rpc_urls in swagger, and use guava for parsing