VB-266: Update Helen to use JRE 11 (#87)
* VB-266: Update Helen to use JRE 11

This patch updates Helen to use JRE 11 runtime, to pave the way for
using JDK 11 for Helen - to be fulfilled in subsequent patch(es).

Also, this patch explicitly states the version of Javassist instead of
letting Maven decide at compile time (inferred from transitive
dependencies).

The reason for the Javassist change is because there are some changes
to what kind of reflective access is allowed in Java 11. The downstream
effect is that Hibernate may not be able to invoke constructors to
create object instances.